Add basic systems and system manager class
Added new oki::System and oki::SystemManager as a very simple
default implementation of systems (the S in ECS).

Note that it is not highly optimized. In fact, it's rather slow.
That said, other parts of the library do not rely on this feature
and end users are free to switch it out as wanted.